gerrit: handle errors with retry
When we get an error from Gerrit, go ahead and retry it every 10
seconds.  This handles issues with Gerrit itself.  For errors in
JSON parsing, continue to abort but give more output in the JSON
failure.

For the abandon operation specifically, Gerrit can return a status
409 (Conflict) when the abandon is not successful.  Do not retry on
this nor crash but just record it as happening.

Signed-off-by: Patrick Williams <patrick@stwcx.xyz>