scripts: download rejected_patterns.yaml from openbmc-tools
The `tof-voters` subtool in openbmc-tools contains a YAML file of
file and repo patterns which are considered "company specific" and
therefore do not count for TOF votes.  We will use this same information
to create a filter of commits that should be rejected from the Discord
"commits to be reviewed" notification.

Add a script which downloads this file and puts it into `config`.